Final committ
Changed register button type to "reg-submit" to avoid an internal server error when clicking.
To avoid server error, now button does not respond to being clicked.
No error messages show up / nothing responds after button is clicked, but information is entered in fine.
Possibly due to using the mock test_user?
But when removed this does not affect the outcome.
Final decision was to leave button type as "submit" and go with the internal server error after learning you can not change the submit type of a button.